+++ .github/release-notes.md
@@ -1,3 +1,3 @@
 ## Changes:
 
--
+- Print debug logs with `-b`, `--debug-log`.